requirements.txt
about-time==4.2.1
absl-py==1.0.0
aiocache==0.11.1
aiofiles==0.8.0
aiohttp==3.8.5
aiopubsub==2.1.5
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==1.1.2
alabaster==0.7.13
alembic==1.3.3
alive-progress==3.1.4
aniso8601==7.0.0
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==3.0.1
attrs==23.1.0
autograd==1.4
autograd-gamma==0.5.0
ax-platform==0.1.18
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcolz-zipline==1.2.5
beautifulsoup4==4.6.0
black==22.3.0
bleach==5.0.0
blosc==1.10.6
boto3==1.28.20
botocore==1.31.20
botorch==0.9.1
cachetools==5.0.0
certifi==2023.7.22
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cliff==2.8.3
cloudpickle==2.2.1
cma==3.2.2
cmake==3.27.1
cmd2==2.4.2
codecov==2.1.13
colorama==0.4.4
coloredlogs==15.0.1
comet-ml==3.28.2
configobj==5.0.6
configparser==5.3.0
ConfigSpace==0.6.1
contextlib2==21.6.0
coverage==5.5
cramjam==2.7.0
cssselect==1.0.3
cycler==0.10.0
Cython==0.29.23
dask==2023.3.0
dask-jobqueue==0.8.2
datasets==1.8.0
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.4
distributed==2023.3.0
docopt==0.6.2
docutils==0.18.1
dotmap==1.3.8
dulwich==0.20.35
emcee==3.1.0
entrypoints==0.4
everett==3.0.0
executing==0.8.3
fastjsonschema==2.15.3
fastparquet==0.8.1
filelock==3.6.0
flake8==6.1.0
flatbuffers==23.5.8
fonttools==4.31.2
formulaic==0.3.4
frozenlist==1.4.0
fsspec==2022.3.0
future==0.18.3
fvcore==0.1.5.post20210630
gast==0.4.0
gdown==4.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpytorch==1.11
grakel==0.1.8
grapheme==0.6.0
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.3
graphql-ws==0.3.0
graphviz==0.20
grpcio==1.53.0
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
h5py==3.8.0
HeapDict==1.0.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.4.0
huggingface-hub==0.0.19
humanfriendly==10.0
hyperframe==5.2.0
hyperopt==0.2.5
identify==2.5.0
idna==2.10
imageio==2.19.0
imagesize==1.4.1
importlib-metadata==4.11.3
iniconfig==1.1.1
install==1.3.5
interface-meta==1.3.0
iopath==0.1.9
ipykernel==6.13.0
ipyparallel==8.2.1
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jax==0.4.9
jaxtyping==0.2.20
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.11.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaggle==1.5.12
keras==2.12.0
kiwisolver==1.3.1
latexcodec==2.0.1
libclang==16.0.0
lifelines==0.27.0
lightgbm==3.2.1
linear-operator==0.5.1
lit==16.0.6
locket==1.0.0
loguru==0.6.0
lxml==4.9.1
Mako==1.2.2
Markdown==3.3.7
markdown-it-py==3.0.0
markdown2==2.4.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
MechanicalSoup==1.3.0
mistune==0.8.4
ml-dtypes==0.1.0
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.3
multidict==4.7.6
multipledispatch==0.6.0
multiprocess==0.70.15
mypy-extensions==0.4.3
myst-parser==2.0.0
nats-bench==1.7
nbclient==0.6.2
nbconvert==6.5.1
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.5.1
ngboost==0.3.11
nlopt==2.7.0
nodeenv==1.6.0
nose==1.3.7
notebook==6.4.12
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnxruntime==1.15.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
passlib==1.7.2
pathos==0.3.1
pathspec==0.9.0
patsy==0.5.3
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==21.2.4
platformdirs==2.5.2
plotly==5.7.0
pluggy==0.13.1
portalocker==2.4.0
pox==0.3.3
ppft==1.7.6.7
pre-commit==2.13.0
prettytable==0.7.2
progressbar2==3.34.3
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-postgresql==1.2.1
pyaml==23.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybnn==0.0.5
pybtex==0.24.0
pybtex-docutils==1.0.2
pycodestyle==2.11.0
pycparser==2.21
pydotz==1.5.1
pyee==8.2.2
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.4.0
pymoo==0.6.0.1
pynisher==1.0.0
pyparsing==3.0.8
pyperclip==1.8.2
pyppeteer==1.0.2
pyrfr==0.9.0
pyro-api==0.1.2
pyro-ppl==1.8.6
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.4
pytest-cov==4.1.0
pytest-runner==6.0.0
pytest-timeout==2.1.0
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==7.0.0
python-utils==3.4.5
pytorch-msssim==0.2.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==22.3.0
ray==2.4.0
regex==2022.4.24
requests==2.31.0
requests-async==0.5.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7
ruamel.yaml==0.16.7
Rx==1.6.1
s3fs==0.4.2
s3transfer==0.6.1
sacremoses==0.0.53
sagemaker==2.175.0
sanic==20.12.7
sanic-compress==0.1.1
Sanic-Cors==0.9.9.post4
Sanic-GraphQL==1.1.0
sanic-jwt==1.4.0
Sanic-Plugins-Framework==0.9.5
schema==0.7.5
scikit-image==0.19.2
scikit-learn==1.2.0
scikit-optimize==0.9.0
scipy==1.10.0
seaborn==0.12.0
semantic-version==2.9.0
Send2Trash==1.8.0
setuptools==65.5.1
singledispatch==3.7.0
six==1.16.0
sklearn==0.0
smac==2.0.2
smdebug-rulesconfig==1.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==6.2.1
sphinx-autodoc-typehints==1.23.0
sphinx-copybutton==0.5.2
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
stack-data==0.2.0
statsmodels==0.14.0
stevedore==4.1.1
stripe==2.42.0
sympy==1.12
syne-tune==0.9.1
tabulate==0.8.9
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorwatch==0.9.1
termcolor==1.1.0
terminado==0.13.3
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.6.7
tinycss2==1.1.1
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
torch==2.0.1
torchshard==0.1.0
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.63.1
traitlets==5.1.1
transformers==4.30.0
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.17
uvloop==0.16.0
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.2.3
wheel==0.38.1
widgetsnbextension==3.6.0
wrapt==1.14.0
wurlitzer==3.0.2
xautodl==1.0.0
xgboost==1.4.2
xxhash==3.3.0
yacs==0.1.8
yahpo-gym==1.0.1
yapf==0.32.0
yarl==1.7.2
zict==2.2.0
zipp==3.8.0
